MAINT: Move _FunctionWrapper to scipy._lib._util
