+++ scipy/integrate/_quad_vec.py
@@ -6,7 +6,7 @@ import functools
 
 import numpy as np
 
-from scipy._lib._util import MapWrapper
+from scipy._lib._util import MapWrapper, _FunctionWrapper
 
 
 class LRUDict(collections.OrderedDict):
@@ -99,18 +99,6 @@ class _Bunch:
     def __repr__(self):
         return "_Bunch({})".format(", ".join("{}={}".format(k, repr(self.__dict__[k]))
                                              for k in self.__keys))
-
-
-class _FunctionWrapper:
-    """
-    Object to wrap user function, allowing picklability
-    """
-    def __init__(self, f, args):
-        self.f = f
-        self.args = args
-
-    def __call__(self, x):
-        return self.f(x, *self.args)
 
 
 def quad_vec(f, a, b, epsabs=1e-200, epsrel=1e-8, norm='2', cache_size=100e6, limit=10000,
